+++ documentation/docs/user/releasenotes.md
@@ -2,6 +2,7 @@
 
 ## Version 0.2.0
 - Added Skip action to skip the first 'n' data items of a stream.
+- Added Limit action to filter data items from the stream after the first 'n' data items have been processed.
 
 ## Version 0.1.0
 - First stable release with core features.